- Split EpicGraphicView into EditableGraphicView - Remove CommonZoom from Code - Promote Send Graphic View to EditableGraphicView
