+++ src/urh/controller/SendRecvDialogController.py
@@ -120,9 +120,9 @@ class SendRecvDialogController(QDialog):
         if mode == Mode.receive or mode == Mode.send:
             self.scene_creator = LiveSceneManager(np.array([]), parent=self) # set really in on_device_started
         else:
-            self.scene_creator = FFTSceneManager(parent=self, graphic_view=self.ui.graphicsView)
+            self.scene_creator = FFTSceneManager(parent=self, graphic_view=self.ui.graphicsViewReceive)
 
-        self.ui.graphicsView.setScene(self.scene_creator.scene)
+        self.ui.graphicsViewReceive.setScene(self.scene_creator.scene)
 
         ipRange = "(?:[0-1]?[0-9]?[0-9]|2[0-4][0-9]|25[0-5])"
         ipRegex = QRegExp("^" + ipRange
@@ -152,11 +152,11 @@ class SendRecvDialogController(QDialog):
         self.ui.cbDevice.currentIndexChanged.connect(self.on_selected_device_changed)
         self.ui.spinBoxNRepeat.editingFinished.connect(self.on_nrepeat_changed)
         self.ui.sliderYscale.valueChanged.connect(self.on_slideyscale_value_changed)
-        self.ui.graphicsView.freq_clicked.connect(self.on_graphics_view_freq_clicked)
+        self.ui.graphicsViewReceive.freq_clicked.connect(self.on_graphics_view_freq_clicked)
         self.ui.btnLockBWSR.clicked.connect(self.on_btn_lock_bw_sr_clicked)
 
-        self.ui.graphicsView.zoomed.connect(self.handle_signal_zoomed_or_scrolled)
-        self.ui.graphicsView.horizontalScrollBar().valueChanged.connect(self.handle_signal_zoomed_or_scrolled)
+        self.ui.graphicsViewReceive.zoomed.connect(self.handle_signal_zoomed_or_scrolled)
+        self.ui.graphicsViewReceive.horizontalScrollBar().valueChanged.connect(self.handle_signal_zoomed_or_scrolled)
 
     def __create_device_connects(self):
         self.device.stopped.connect(self.on_device_stopped)
@@ -222,7 +222,7 @@ class SendRecvDialogController(QDialog):
         del self.scene_creator.plot_data
         if self.mode == Mode.receive or self.mode == Mode.send:
             self.scene_creator = LiveSceneManager(np.array([]), parent=self)
-        self.ui.graphicsView.setScene(self.scene_creator.scene)
+        self.ui.graphicsViewReceive.setScene(self.scene_creator.scene)
         self.ui.lineEditIP.setVisible(dev_name == "USRP")
         self.ui.labelIP.setVisible(dev_name == "USRP")
 
@@ -258,7 +258,7 @@ class SendRecvDialogController(QDialog):
 
     @pyqtSlot()
     def on_device_stopped(self):
-        self.ui.graphicsView.capturing_data = False
+        self.ui.graphicsViewReceive.capturing_data = False
         if self.mode == Mode.send:
             self.ui.btnStart.setIcon(QIcon.fromTheme("media-playback-start"))
             self.ui.btnStart.setToolTip("Start sending")
@@ -286,7 +286,7 @@ class SendRecvDialogController(QDialog):
 
         self.ui.txtEditErrors.clear()
         self.scene_creator.set_text("Waiting for device..")
-        self.ui.graphicsView.capturing_data = True
+        self.ui.graphicsViewReceive.capturing_data = True
         self.ui.btnSave.setEnabled(False)
 
         if self.mode == Mode.send:
@@ -353,7 +353,7 @@ class SendRecvDialogController(QDialog):
             return
 
         if self.mode == Mode.send or self.mode == Mode.receive:
-            self.ui.graphicsView.horizontalScrollBar().blockSignals(True)
+            self.ui.graphicsViewReceive.horizontalScrollBar().blockSignals(True)
             self.scene_creator.end = self.device.current_index
         elif self.mode == Mode.spectrum:
             x, y = self.device.spectrum
@@ -365,7 +365,7 @@ class SendRecvDialogController(QDialog):
 
         self.scene_creator.init_scene()
         self.scene_creator.show_full_scene()
-        self.ui.graphicsView.update()
+        self.ui.graphicsViewReceive.update()
 
     def __restart_device_thread(self):
         self.device.stop("Restarting with new port")
@@ -384,7 +384,7 @@ class SendRecvDialogController(QDialog):
             self.device.current_index = 0
             self.scene_creator.clear_path()
             self.device.current_iteration = 0
-            self.ui.graphicsView.repaint()
+            self.ui.graphicsViewReceive.repaint()
             self.ui.lSamplesCaptured.setText("0")
             self.ui.lSignalSize.setText("0")
             self.ui.lTime.setText("0")
@@ -440,9 +440,9 @@ class SendRecvDialogController(QDialog):
         event.accept()
 
     def handle_signal_zoomed_or_scrolled(self):
-        if not self.ui.graphicsView.capturing_data:
-            x1 = self.ui.graphicsView.view_rect().x()
-            x2 = x1 + self.ui.graphicsView.view_rect().width()
+        if not self.ui.graphicsViewReceive.capturing_data:
+            x1 = self.ui.graphicsViewReceive.view_rect().x()
+            x2 = x1 + self.ui.graphicsViewReceive.view_rect().width()
             self.scene_creator.show_scene_section(x1, x2)
 
     def on_btn_lock_bw_sr_clicked(self):
@@ -459,10 +459,10 @@ class SendRecvDialogController(QDialog):
         # Scale Up = Top Half, Scale Down = Lower Half
         middle = int((self.ui.sliderYscale.maximum() + 1 - self.ui.sliderYscale.minimum()) / 2)
         scale_up = new_value >= middle
-        current_factor = self.ui.graphicsView.sceneRect().height() / self.ui.graphicsView.view_rect().height()
+        current_factor = self.ui.graphicsViewReceive.sceneRect().height() / self.ui.graphicsViewReceive.view_rect().height()
         scale_factor = (new_value + 1 - middle) / current_factor if scale_up else current_factor / new_value
         if scale_factor > 0:
-            self.ui.graphicsView.scale(1, scale_factor)
+            self.ui.graphicsViewReceive.scale(1, scale_factor)
 
     @pyqtSlot(float)
     def on_graphics_view_freq_clicked(self, freq: float):
