+++ src/urh/ui/ui_main.py
@@ -151,10 +151,6 @@ class Ui_MainWindow(object):
         icon = QtGui.QIcon.fromTheme("edit-redo")
         self.actionRedo.setIcon(icon)
         self.actionRedo.setObjectName("actionRedo")
-        self.actionCommon_Zoom = QtWidgets.QAction(MainWindow)
-        self.actionCommon_Zoom.setCheckable(True)
-        self.actionCommon_Zoom.setChecked(False)
-        self.actionCommon_Zoom.setObjectName("actionCommon_Zoom")
         self.actionShow_Confirm_Close_Dialog = QtWidgets.QAction(MainWindow)
         self.actionShow_Confirm_Close_Dialog.setCheckable(True)
         self.actionShow_Confirm_Close_Dialog.setChecked(False)
@@ -302,7 +298,6 @@ class Ui_MainWindow(object):
         self.actionAuto_Fit_Y.setText(_translate("MainWindow", "&Auto Fit Y"))
         self.actionUndo.setText(_translate("MainWindow", "&Undo"))
         self.actionRedo.setText(_translate("MainWindow", "&Redo"))
-        self.actionCommon_Zoom.setText(_translate("MainWindow", "&Common Zoom"))
         self.actionShow_Confirm_Close_Dialog.setText(_translate("MainWindow", "&Show Confirm Close Dialog"))
         self.actionTest.setText(_translate("MainWindow", "test"))
         self.actionHold_Shift_to_Drag.setText(_translate("MainWindow", "&Hold Shift to Drag"))
