+++ src/urh/ui/ui_send_recv.py
@@ -13,6 +13,36 @@ class Ui_SendRecvDialog(object):
         SendRecvDialog.resize(868, 628)
         self.gridLayout_3 = QtWidgets.QGridLayout(SendRecvDialog)
         self.gridLayout_3.setObjectName("gridLayout_3")
+        self.sliderYscale = QtWidgets.QSlider(SendRecvDialog)
+        self.sliderYscale.setMinimum(1)
+        self.sliderYscale.setMaximum(10)
+        self.sliderYscale.setProperty("value", 5)
+        self.sliderYscale.setOrientation(QtCore.Qt.Vertical)
+        self.sliderYscale.setObjectName("sliderYscale")
+        self.gridLayout_3.addWidget(self.sliderYscale, 1, 2, 1, 1)
+        self.stackedWidget = QtWidgets.QStackedWidget(SendRecvDialog)
+        self.stackedWidget.setObjectName("stackedWidget")
+        self.page_receive = QtWidgets.QWidget()
+        self.page_receive.setObjectName("page_receive")
+        self.verticalLayout_2 = QtWidgets.QVBoxLayout(self.page_receive)
+        self.verticalLayout_2.setContentsMargins(0, 0, 0, 0)
+        self.verticalLayout_2.setObjectName("verticalLayout_2")
+        self.graphicsViewReceive = LiveGraphicView(self.page_receive)
+        self.graphicsViewReceive.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOff)
+        self.graphicsViewReceive.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
+        self.graphicsViewReceive.setObjectName("graphicsViewReceive")
+        self.verticalLayout_2.addWidget(self.graphicsViewReceive)
+        self.stackedWidget.addWidget(self.page_receive)
+        self.page_send = QtWidgets.QWidget()
+        self.page_send.setObjectName("page_send")
+        self.verticalLayout_3 = QtWidgets.QVBoxLayout(self.page_send)
+        self.verticalLayout_3.setContentsMargins(0, 0, 0, 0)
+        self.verticalLayout_3.setObjectName("verticalLayout_3")
+        self.graphicsViewSend = EditableGraphicView(self.page_send)
+        self.graphicsViewSend.setObjectName("graphicsViewSend")
+        self.verticalLayout_3.addWidget(self.graphicsViewSend)
+        self.stackedWidget.addWidget(self.page_send)
+        self.gridLayout_3.addWidget(self.stackedWidget, 0, 1, 3, 1)
         self.verticalLayout = QtWidgets.QVBoxLayout()
         self.verticalLayout.setObjectName("verticalLayout")
         self.gridLayout = QtWidgets.QGridLayout()
@@ -221,26 +251,15 @@ class Ui_SendRecvDialog(object):
         self.txtEditErrors.setObjectName("txtEditErrors")
         self.verticalLayout.addWidget(self.txtEditErrors)
         self.gridLayout_3.addLayout(self.verticalLayout, 0, 0, 2, 1)
-        self.graphicsView = LiveGraphicView(SendRecvDialog)
-        self.graphicsView.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOff)
-        self.graphicsView.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
-        self.graphicsView.setObjectName("graphicsView")
-        self.gridLayout_3.addWidget(self.graphicsView, 0, 1, 2, 1)
         self.label_6 = QtWidgets.QLabel(SendRecvDialog)
         self.label_6.setObjectName("label_6")
         self.gridLayout_3.addWidget(self.label_6, 0, 2, 1, 1)
-        self.sliderYscale = QtWidgets.QSlider(SendRecvDialog)
-        self.sliderYscale.setMinimum(1)
-        self.sliderYscale.setMaximum(10)
-        self.sliderYscale.setProperty("value", 5)
-        self.sliderYscale.setOrientation(QtCore.Qt.Vertical)
-        self.sliderYscale.setObjectName("sliderYscale")
-        self.gridLayout_3.addWidget(self.sliderYscale, 1, 2, 1, 1)
         self.gridLayout_3.setColumnStretch(0, 1)
         self.gridLayout_3.setColumnStretch(1, 5)
         self.gridLayout_3.setColumnStretch(2, 1)
 
         self.retranslateUi(SendRecvDialog)
+        self.stackedWidget.setCurrentIndex(1)
         QtCore.QMetaObject.connectSlotsByName(SendRecvDialog)
         SendRecvDialog.setTabOrder(self.cbDevice, self.lineEditIP)
         SendRecvDialog.setTabOrder(self.lineEditIP, self.spinBoxFreq)
@@ -252,8 +271,7 @@ class Ui_SendRecvDialog(object):
         SendRecvDialog.setTabOrder(self.btnStart, self.btnStop)
         SendRecvDialog.setTabOrder(self.btnStop, self.btnSave)
         SendRecvDialog.setTabOrder(self.btnSave, self.btnClear)
-        SendRecvDialog.setTabOrder(self.btnClear, self.graphicsView)
-        SendRecvDialog.setTabOrder(self.graphicsView, self.sliderYscale)
+        SendRecvDialog.setTabOrder(self.btnClear, self.sliderYscale)
         SendRecvDialog.setTabOrder(self.sliderYscale, self.txtEditErrors)
 
     def retranslateUi(self, SendRecvDialog):
@@ -288,5 +306,6 @@ class Ui_SendRecvDialog(object):
         self.label_6.setText(_translate("SendRecvDialog", "Y-Scale"))
 
 from urh.ui.KillerDoubleSpinBox import KillerDoubleSpinBox
+from urh.ui.views.EditableGraphicView import EditableGraphicView
 from urh.ui.views.LiveGraphicView import LiveGraphicView
 from . import urh_rc
